[feat] 직원 등록 시 월급 테이블에 자동 등록되는 기능 구현

+++ src/main/java/com/bubblebubble/hr/payment/dto/PayrollLedgerRequestDTO.java
@@ -14,7 +14,7 @@ import lombok.NoArgsConstructor;
 @Data
 public class PayrollLedgerRequestDTO {
     Long paymentNo;
-    List<Long> empNo;
+    List<Integer> empNo;
     Long payrollNo;
     String name;
     @DateTimeFormat(pattern = "yyyy-MM-dd")
